+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+* Fix the bug of the table constructor
+* Add auxiliary API for reference(`ref`/`unRef`)
+
+## 0.0.2
+* Add Lua userdata support
+* Fix lexical analysis BUG
+
 ## 0.0.1
 * A full lua virtual machine
 * Support some standard libraries, e.g. String, Math, etc.